Add existence check in delete, info and restore commands

+++ src/main/java/ukitinu/markovwords/cmd/DeleteCmd.java
@@ -26,6 +26,10 @@ public class DeleteCmd implements Callable<Integer> {
 
     @Override
     public Integer call() {
+        if (!repo.exists(name)) {
+            printStream.println("Given dictionary does not exists: " + name);
+            return 1;
+        }
         try {
             repo.delete(name, permanent);
             printStream.println(permanent ? "Dictionary deleted permanently: " + name : "Dictionary deleted: " + name);
